pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.135.1"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
pydantic-settings = "^2.13.1"
loguru = "^0.7.3"
numpy = "^2.4.3"
scipy = "^1.17.1"
openai = "^2.28.0"
httpx = "^0.28.1"
python-dotenv = "^1.2.2"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
